+++ lib/xliff/entry.rb
@@ -105,6 +105,7 @@ module Xliff
       raise 'Entry XML is nil' if xml.nil?
       raise "Invalid Entry XML – must be a nokogiri object, got `#{xml.class}`" unless xml.is_a? Nokogiri::XML::Element
       raise 'Invalid Entry XML – the root node must be `<trans-unit>`' if xml.name != 'trans-unit'
+      raise 'Invalid Entry XML – `<trans-unit>` is missing the required `id` attribute' if xml['id'].nil?
       raise 'Invalid Entry XML – `<trans-unit>` is missing a `<source>` element' if xml.at('source').nil?
     end
   end